+++ tests/test_utils.py
@@ -94,40 +94,30 @@ def test_can_set_cached_property():
     assert a._prop == "value"
 
 
-def test_can_invalidate_cached_property():
-    foo = []
+def test_invalidate_cached_property():
+    accessed = 0
 
     class A:
+        @utils.cached_property
         def prop(self):
-            foo.append(42)
+            nonlocal accessed
+            accessed += 1
             return 42
-
-        prop = utils.cached_property(prop)
 
     a = A()
     p = a.prop
     q = a.prop
     assert p == q == 42
-    assert foo == [42]
+    assert accessed == 1
 
-    utils.invalidate_cached_property(a, "prop")
+    a.prop = 16
+    assert a.prop == 16
+    assert accessed == 1
+
+    del a.prop
     r = a.prop
     assert r == 42
-    assert foo == [42, 42]
-
-    s = a.prop
-    assert s == 42
-    assert foo == [42, 42]
-
-
-def test_invalidate_cached_property_on_non_property():
-    class A:
-        def __init__(self):
-            self.prop = 42
-
-    a = A()
-    with pytest.raises(TypeError):
-        utils.invalidate_cached_property(a, "prop")
+    assert accessed == 2
 
 
 def test_inspect_treats_cached_property_as_property():